\name{brad302}
\alias{brad302}
\docType{data}
\title{Hypothesized random normal N(3, 2) deviates
}
\description{Data hypothesized to come from a normal distribution with mean = 3 and standard deviation = 2 to be evaluated with 1-sample coverage test.
}
\usage{data(brad302)}
\format{
  A data frame with 5 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{The random normal deviate.}
    \item{\code{fx}}{Transformed cumulative probabilities of x from a N(3, 2) distribution.}
  }
}
\source{
Bradley, J.V. 1968. Distribution free statistical tests. \emph{Prentice Hall}, Inc. pp. 301--302.
}

\examples{

print(brad302)
}
\keyword{datasets}
