\name{paired1}
\alias{paired1}
\docType{data}
\title{Example of paired data set up in a blocked structure
}
\description{Paired data structured as a complete randomized block design with 10 blocks for pairs, 2 treatment groups, and 20 responses.
}
\usage{data(paired1)}
\format{
  A data frame with 20 observations on the following 3 variables.
  \describe{
    \item{\code{group}}{The 2 treatment groups.}
    \item{\code{pair}}{The 10 blocks.}
    \item{\code{response}}{The 20 hypothetical responses.}
  }
}

\examples{
str(paired1)
}
\keyword{datasets}
