\name{paired2}
\alias{paired2}
\docType{data}
\title{Paired data structure
}
\description{Paired data structure with 2 columns for the first and second responses for 10 pairs of values.
}
\usage{data(paired2)}
\format{
  A data frame with 10 observations on the following 2 variables.
  \describe{
    \item{\code{first}}{Responses for treatment 1.}
    \item{\code{second}}{Responses for treatment 2.}
  }
}
\details{This is an alternative data representation for the data in paired1.
}

\examples{
str(paired2)
}
\keyword{datasets}
